requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
alembic
pydantic-settings
python-jose
python-multipart
